initial cleanup and restructuring - standardizes express app structure with existing curris projects - pro-coding fixes (formatting, spacing, semi-colons, use strict) - adds docstrings
