+++ config.js
@@ -1,0 +1,7 @@
+"use strict";
+
+const PORT = process.env.PORT || 3000;
+
+module.exports = {
+  PORT,
+};
