+++ server.js
@@ -1,0 +1,8 @@
+"use strict";
+
+const app = require("./app");
+const { PORT } = require("./config");
+
+app.listen(PORT, function () {
+  console.log(`Started on http://localhost:${PORT}`);
+});
